[MOVE] PhoneVerification 관련 repository 클래스를 phone.verification 패키지 하위로 이동
