+++ src/main/java/sopt/makers/authentication/database/rdb/repository/phone/verification/PhoneVerificationRemover.java
@@ -1,4 +1,4 @@
-package sopt.makers.authentication.database.rdb.repository.auth;
+package sopt.makers.authentication.database.rdb.repository.phone.verification;
 
 import sopt.makers.authentication.domain.auth.PhoneVerification;
