+++ src/main/java/sopt/makers/authentication/database/rdb/repository/phone/verification/PhoneVerificationRepositoryImpl.java
@@ -1,9 +1,6 @@
-package sopt.makers.authentication.database;
+package sopt.makers.authentication.database.rdb.repository.phone.verification;
 
 import sopt.makers.authentication.database.rdb.entity.PhoneVerificationEntity;
-import sopt.makers.authentication.database.rdb.repository.auth.PhoneVerificationRegister;
-import sopt.makers.authentication.database.rdb.repository.auth.PhoneVerificationRemover;
-import sopt.makers.authentication.database.rdb.repository.auth.PhoneVerificationRetriever;
 import sopt.makers.authentication.domain.auth.PhoneVerification;
 import sopt.makers.authentication.usecase.auth.port.out.PhoneVerificationRepository;
