+++ src/main/java/sopt/makers/authentication/database/rdb/repository/phone/verification/PhoneVerificationRetriever.java
@@ -1,4 +1,4 @@
-package sopt.makers.authentication.database.rdb.repository.auth;
+package sopt.makers.authentication.database.rdb.repository.phone.verification;
 
 import sopt.makers.authentication.database.rdb.entity.PhoneVerificationEntity;
 import sopt.makers.authentication.domain.auth.PhoneVerification;
